refactor: Update BusinessSearchList component (#1275)
Refactor the BusinessSearchList component to improve the handling of term value changes. The code now checks if the length of the new term value is less than 2 and the previous term value was at least 2 characters long. Additionally, the new term value is converted to lowercase before performing the search

Co-authored-by: YisusJe <jesus@ordering.co>